# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: 文旅消费券积分平台 API
  version: 1.0.0
  description: |
    消费者端、商户端、管理端统一 API 设计（MVP）。
    统一响应：`{ code, message, traceId, data }`。
servers:
  - url: https://api.example.com
    description: production
  - url: http://localhost:3000
    description: local
...
paths:
...
  /api/ticket/list:
    get:
      tags: [Ticket]
      summary: 我的票据列表
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: status
          schema:
            type: integer
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: pageSize
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: ok
...
  /api/blindbox/open:
    post:
      tags: [BlindBox]
      summary: 打开盲盒
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ticketId]
              properties:
                ticketId:
                  type: integer
                  format: int64
                blindBoxId:
                  type: integer
                  format: int64
      responses:
        '200':
          description: 返回奖品与券信息
  /api/blindbox/history:
    get:
      tags: [BlindBox]
      summary: 开奖历史
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: pageSize
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: ok
  /api/blindbox/config:
    get:
      tags: [BlindBox]
      summary: 当前盲盒配置（不含敏感概率细节时可裁剪）
      security:
        - bearerAuth: []
      responses:
        '200':
          description: ok
...
  /api/point/goods:
    get:
      tags: [Point]
      summary: 积分商城商品
      security:
        - bearerAuth: []
      responses:
        '200':
          description: ok
...
  /admin/dashboard:
    get:
      tags: [Admin]
      summary: 运营仪表盘
      security:
        - bearerAuth: []
      responses:
        '200':
          description: ok
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...